Introduce a generic Meter base class
